Support protobuf input for eval
